backend/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.1.*
flask-marshmallow==1.2.*
marshmallow-sqlalchemy==1.1.*
marshmallow==3.26.*
Flask-CORS==5.0.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
pytest==8.3.*
